Classification-of-Stroke-Blood-Clot-Origin/requirements.txt
openslide-python==1.2.0
torch
torchvision
pytorch-lightning
torchmetrics
tqdm==4.65.0
pandas==1.5.3
matplotlib==3.7.1
seaborn==0.12.2
opencv-python==4.7.0.72
pillow==9.4.0
Flask==2.2.2
mysql-connector-python